organize components into folders
